structs.h: fix int typedefs

+++ src/structs.h
@@ -1,11 +1,11 @@
-typedef unsigned short int u8;
-typedef unsigned int u16;
-typedef unsigned long u32;
-typedef unsigned long long u64;
-typedef short int i8;
-typedef int i16;
-typedef long i32;
-typedef long long i64;
+typedef uint8_t u8;
+typedef uint16_t u16;
+typedef uint32_t u32;
+typedef uint64_t u64;
+typedef int8_t i8;
+typedef int16_t i16;
+typedef int32_t i32;
+typedef int64_t i64;
 
 
 //Superblock Flags
